nxp-imx6: put x11 as the first factor in DISTRO_FEATURES
RallyID: US107623

In freescale's SDK, BACKEND is either x11 or wayland. But in WRLinux,
when bitbake wrlinux-image-glibc-std-sato to build a x11 BACKEND graphic
system, x11 and wayland are in DISTRO_FEATURES in the meantime.
So x11 should be treated as the first factor in judging BACKEND, or else
these app will be configured as a wayland-backend when they check "wayland"
in the DISTRO_FEATURES.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>