nxp-imx6.conf: Allow multi machine builds with oe-core
When building more than one target which use u-boot each bsp must
define a MACHINE specific bootloader preference if it needs to use a
specific boot loader as is the case with the nxp-imx6 design.

If not defined, a build error will occur with a conflict with the
u-boot found in oe-core.

[ Issue: LINUXEXEC-3159 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>