nxp-imx6: remove setting WRL_RECIPES from templates
Issue: LINUXEXEC-26075

Some customers want access to use recipes that are included in
WRL_RECIPES but they do not use templates. By moving the updates to
WRL_RECIPES outside of the feature and default templates we can
simplify this usecase.

We move the changes to WRL_RECIPES to wrlinux-recipelists.conf.

Signed-off-by: Sakib Sajal <sakib.sajal@windriver.com>